+++ src/get_text/wezterm.rs
@@ -3,7 +3,9 @@ use std::{env, process::Command};
 use super::find_command_output;
 
 pub fn get_text(cmd: &str, depth: usize) -> Option<String> {
-    env::var("WEZTERM_EXECUTABLE").ok()?;
+    if &env::var("TERM_PROGRAM").ok()? != "WezTerm" {
+        return None;
+    }
 
     let output = Command::new("wezterm")
         .args(["cli", "get-text", "--start-line", &format!("-{depth}")])